requirements.txt
fastapi==0.110.0
pydantic==2.6.4
uvicorn==0.29.0
asyncpg==0.29.0
bcrypt==4.1.2
sqlalchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.1
python-jose==3.3.0
alembic-postgresql-enum==1.1.2
pytest==8.0.2
faker==24.1.0
httpx==0.27.0
PyJWT==2.8.0
pytest-asyncio==0.23.5
psycopg2==2.9.9